refactor: extract utilities into focused modules and fix force sync issues
BREAKING CHANGES:
- Modularized src/utils.js into domain-specific utility modules
- Fixed force sync option to properly bypass cache checks

Features:
- Created src/utils/ directory with 6 specialized modules:
  * time.js - Duration formatting and parsing utilities
  * network.js - HTTP agents, token normalization, retry logic
  * concurrency.js - Semaphore and RateLimiter classes
  * data.js - Safe parsing functions (int, float, boolean)
  * debug.js - Debug and diagnostic utilities
  * api-testing.js - API connection testing utilities
- Added utils/index.js for convenient re-exports

Fixes:
- Fixed --force sync option bypassing cache checks correctly
- Fixed force sync processing completed books that were previously skipped
- Fixed progress regression protection respecting force flag
- Consolidated duplicate HTTP agent creation logic
- Moved parseDurationString from hardcover-extractor to time utils
- Removed deprecated calculateCurrentPage/calculateCurrentSeconds functions

Technical improvements:
- Better separation of concerns with focused utility modules
- Eliminated code duplication across client classes
- Maintained dependency injection pattern for config values
- Preserved all existing functionality while improving maintainability
- Fixed SyntaxError in hardcover-extractor.js (await import issue)

Updated imports across:
- audiobookshelf-client.js
- hardcover-client.js
- sync-manager.js
- main.js
- config-validator.js
- matching/utils/hardcover-extractor.js

Documentation:
- Updated Architecture-Overview.md to reflect new modular utility structure
