+++ src/main.js
@@ -7,7 +7,8 @@ import { SyncManager } from './sync-manager.js';
 import { AudiobookshelfClient } from './audiobookshelf-client.js';
 import { HardcoverClient } from './hardcover-client.js';
 import { BookCache } from './book-cache.js';
-import { dumpFailedSyncBooks, testApiConnections } from './utils.js';
+import { dumpFailedSyncBooks } from './utils/debug.js';
+import { testApiConnections } from './utils/api-testing.js';
 import { currentVersion } from './version.js';
 import cron from 'node-cron';
 import logger from './logger.js';
