+++ src/utils/index.js
@@ -1,0 +1,35 @@
+/**
+ * Utilities index - Re-exports all utility functions for easier imports
+ *
+ * This allows importing multiple utilities from different modules with:
+ * import { formatDuration, Semaphore, normalizeApiToken } from './utils/index.js';
+ *
+ * Instead of multiple imports from different files.
+ */
+
+// Time utilities
+export {
+  formatDuration,
+  formatDurationForLogging,
+  parseDurationString,
+  sleep,
+} from './time.js';
+
+// Debug utilities
+export { dumpFailedSyncBooks } from './debug.js';
+
+// Network utilities
+export {
+  normalizeApiToken,
+  createHttpAgent,
+  retryWithBackoff,
+} from './network.js';
+
+// Concurrency utilities
+export { Semaphore, RateLimiter } from './concurrency.js';
+
+// API testing utilities
+export { testApiConnections } from './api-testing.js';
+
+// Data parsing utilities
+export { safeParseInt, safeParseFloat, safeParseBoolean } from './data.js';
